backend/requirements.txt
openai~=1.78.0
uvicorn~=0.34.2
fastapi~=0.115.12
pydantic
sqlalchemy
sqlmodel
langchain_text_splitters
langchain_hyperbrowser
psycopg2-binary
python-dotenv
markdown
groq

